requirements.txt
python-telegram-bot==20.3
gspread==5.7.2
google-auth==2.21.0
Flask==2.3.2
